+++ tests/tables/cmap.rs
@@ -488,6 +488,37 @@ mod format4 {
     }
 
     #[test]
+    fn invalid_offset() {
+        let data = convert(&[
+            UInt16(4), // format
+            UInt16(42), // subtable size
+            UInt16(0), // language ID
+            UInt16(4), // 2 x segCount
+            UInt16(2), // search range
+            UInt16(0), // entry selector
+            UInt16(2), // range shift
+            // End character codes
+            UInt16(69), // char code [0]
+            UInt16(65535), // char code [1]
+            UInt16(0), // reserved
+            // Start character codes
+            UInt16(65), // char code [0]
+            UInt16(65535), // char code [1]
+            // Deltas
+            Int16(0), // delta [0]
+            Int16(1), // delta [1]
+            // Offsets into Glyph index array
+            UInt16(4), // offset [0]
+            UInt16(65535), // offset [1]
+            // Glyph index array
+            UInt16(1), // glyph ID [0]
+        ]);
+
+        let subtable = cmap::Subtable4::parse(&data).unwrap();
+        assert_eq!(subtable.glyph_index(65535), None);
+    }
+
+    #[test]
     fn collect_codepoints() {
         let data = convert(&[
             UInt16(4), // format